# app/docs/swagger.yaml
definitions:
...
  muser.EditUserInfoParams:
    properties:
      avatar_id:
        example: 6
        type: integer
      born:
        example: "1993-06-20"
        type: string
      country_id:
        example: 1
        type: integer
      gender:
        example: 1
        type: integer
      nick_name:
        example: 陈二狗
        type: string
      signature:
        example: emmmmmmmm
        type: string
    type: object
  muser.FeedbackParam:
    properties:
      describe:
        example: 问题描述
        type: string
      phone:
        example: 手机号
        type: string
      pics:
        example: 图片列表
        type: string
      problem:
        example: 遇到的问题
        type: string
    required:
    - describe
    type: object
  muser.LoginParams:
    properties:
      op_token:
        example: 客户端返回的运营商token
        type: string
      operator:
        example: 客户端返回的运营商，CMCC:中国移动通信, CUCC:中国联通通讯, CTCC:中国电信
        type: string
      platform:
        example: 0
        type: integer
      token:
        example: 客户端token
        type: string
    required:
    - op_token
    - operator
    - token
    type: object
  muser.QQLoginParams:
    properties:
      access_token:
        example: 授权token
        type: string
      openid:
        type: string
      platform:
        example: qzone、pengyou或qplus
        type: string
    type: object
  muser.UserInfoResp:
    properties:
      age:
        example: 27
        type: integer
      avatar:
        example: 头像地址
        type: string
      background_img:
        example: 背景图
        type: string
      born:
        example: 出生日期
        type: string
      country:
        example: 0
        type: integer
      gender:
        example: 0
        type: integer
      is_anchor:
        example: 0
        type: integer
      is_attention:
        example: 0
        type: integer
      mobile_num:
        example: 13177656222
        type: integer
      nick_name:
        example: 昵称 陈二狗
        type: string
      signature:
        example: 个性签名
        type: string
      status:
        example: 0
        type: integer
      user_id:
        example: "2009011314521111"
        type: string
      user_type:
        example: 0
        type: integer
    type: object
...
  muser.UserZoneInfoParam:
    properties:
      user_id:
        type: string
    type: object
  muser.WeiboLoginParams:
    properties:
      access_token:
        example: 通行token
        type: string
      uid:
        type: integer
    required:
    - access_token
    - uid
    type: object
  muser.WxLoginParam:
    properties:
      code:
        example: code码
        type: string
    required:
    - code
    type: object
...
  sms.SendSmsCodeParams:
    properties:
      mobile_num:
        type: string
      send_type:
        type: string
    type: object
  sms.SmsCodeLoginParams:
    properties:
      code:
        type: string
      mobile_num:
        type: string
      platform:
        type: integer
    type: object
...
  swag.LoginSwag:
    properties:
      token:
        description: token
        type: string
      user_info:
        $ref: '#/definitions/swag.User'
        description: 用户信息
        type: object
    type: object
  swag.User:
    properties:
      age:
        type: integer
      avatar:
        type: string
      background_img:
        type: string
      born:
        type: string
      channel_id:
        type: integer
      city:
        type: string
      create_at:
        type: integer
      device_type:
        type: integer
      gender:
        type: integer
      id:
        type: integer
      is_anchor:
        type: integer
      last_login_time:
        type: integer
      mobile_num:
        type: integer
      nick_name:
        type: string
      password:
        type: string
      signature:
        type: string
      status:
        type: integer
      title:
        type: string
      update_at:
        type: integer
      user_id:
        type: string
    type: object
  swag.UserZoneInfoResp:
    properties:
      totalAttention:
        description: 关注数
        example: 100
        type: integer
      totalBeLiked:
        description: 被点赞数
        example: 100
        type: integer
      totalCollect:
        description: 收藏的作品数
        example: 100
        type: integer
      totalFans:
        description: 粉丝数
        example: 100
        type: integer
      totalLikes:
        description: 点赞的作品数
        example: 100
        type: integer
      totalPublish:
        description: 发布的作品数
        example: 100
        type: integer
    type: object
  swag.ZoneInfoSwag:
    properties:
      user_info:
        $ref: '#/definitions/muser.UserInfoResp'
        type: object
      zone_info:
        $ref: '#/definitions/swag.UserZoneInfoResp'
        type: object
    type: object
info:
  contact: {}
  description: |-
    ### 一、公共参数说明（此栏参数均为Headers请求头传递）
    | 参数名 | 说明 | 示例 |
    | ------ | :----- | :----- |
    | AppId | AppId(区分Android,iOS,web) | 5EesXF1i |
    | Secret | 服务端下发的secret 通过/api/v1/client/init接口获取 调用该接口时无需传 且 不参与签名 | DnaukFwVILpcewX6 |
    | Timestamp | 请求时间戳 单位：秒 | 1588888888 |
    | Sign | 签名 | 签名 md5签名32位值 |
    | Version | 当前版本 | 1.0.1 |
    ### 二、请求体说明（此栏参数均为POST JSON传递，不可用form-data提交）

    {
    'mobileNum': '13177656222',
    'platform': 0
    }
    ### 三、接口签名生成方式
    签名加密示例:
    params = 请求的url路径(不包含域名与参数) + & + Header头参数以`&`拼接（无需按照字典序，具体看以下栗子） + & + appKey
    appKey由服务端下发 并进行保存
    sign = md5(params) 取md5 32位小写
    如：md5(/api/v1/user/mobile/login&AppId=5EesXF1i&Timestamp=1588888888&Version=1.0.1&Secret=DnaukFwVILpcewX6&RfhHecN9zsNcy19Y)
    appKey为RfhHecN9zsNcy19Y
    ### 四、API错误码文档
    [点击查看](/api/v1/doc)
    ### 五、HTTP状态码说明
    | 状态码 | 说明 |
    | ------ | :----- |
    | 200 | 操作成功 |
    | 400 | 参数错误 |
    | 500 | 内部错误 |
  license: {}
  title: 电竞社区平台（应用服）
  version: "1.0"
paths:
...
  /api/v1/user/edit/info:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 编辑用户信息请求参数
        in: body
        name: EditUserInfoParams
        required: true
        schema:
          $ref: '#/definitions/muser.EditUserInfoParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 编辑用户信息 (ok)
      tags:
      - 账号体系
  /api/v1/user/feedback:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 用户反馈问题 请求参数
        in: body
        name: FeedbackParam
        required: true
        schema:
          $ref: '#/definitions/muser.FeedbackParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 用户反馈问题 (ok)
      tags:
      - 账号体系
  /api/v1/user/info:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/muser.UserInfoResp'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 用户信息 (ok)
      tags:
      - 账号体系
  /api/v1/user/mobile/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 手机号登陆/注册 请求参数
        in: body
        name: LoginParams
        required: true
        schema:
          $ref: '#/definitions/muser.LoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.LoginSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 手机一键注册/登陆 (ok)
      tags:
      - 账号体系
  /api/v1/user/qq/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: QQ登陆/注册 请求参数
        in: body
        name: WeiboLoginParams
        required: true
        schema:
          $ref: '#/definitions/muser.QQLoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.LoginSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: QQ注册/登陆 (ok)
      tags:
      - 账号体系
  /api/v1/user/smscode:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 获取短信验证码请求参数
        in: body
        name: SendSmsCodeParams
        required: true
        schema:
          $ref: '#/definitions/sms.SendSmsCodeParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 获取短信验证码 (ok)
      tags:
      - 账号体系
  /api/v1/user/smscode/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 短信验证码登陆/注册 请求参数
        in: body
        name: SmsCodeLoginParams
        required: true
        schema:
          $ref: '#/definitions/sms.SmsCodeLoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.LoginSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 短信验证码注册/登陆 (ok)
      tags:
      - 账号体系
  /api/v1/user/wechat/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 微信登陆/注册 请求参数
        in: body
        name: WxLoginParam
        required: true
        schema:
          $ref: '#/definitions/muser.WxLoginParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.LoginSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 微信注册/登陆 (ok)
      tags:
      - 账号体系
  /api/v1/user/weibo/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 微博登陆/注册 请求参数
        in: body
        name: WeiboLoginParams
        required: true
        schema:
          $ref: '#/definitions/muser.WeiboLoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.LoginSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 微博注册/登陆 (ok)
      tags:
      - 账号体系
  /api/v1/user/zone/info:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: string
      - description: 个人空间用户信息 请求参数
        in: body
        name: UserZoneInfoParam
        required: true
        schema:
          $ref: '#/definitions/muser.UserZoneInfoParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ZoneInfoSwag'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 个人空间用户信息 (ok)
      tags:
      - 账号体系
...
swagger: "2.0"